chore(core): Rename entry point to main.ts
