+++ examples/symbolPlugin/src/main.ts
