+++ packages/core/package.json
@@ -13,8 +13,8 @@
     "type-safety",
     "typescript"
   ],
-  "main": "./dist/index.js",
-  "types": "./dist/index.d.ts",
+  "main": "./dist/main.js",
+  "types": "./dist/main.d.ts",
   "files": [
     "dist/",
     "src/"
